requirements.txt
# --- Core Framework ---
Flask==3.0.3

# --- Database / ORM ---
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1

# --- Database Migrations ---
Alembic==1.13.2
Flask-Migrate==4.1.0

# --- Validation / Serialization ---
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0

# --- Authentication / Security ---
PyJWT==2.9.0
passlib[bcrypt]==1.7.4    # хэширование паролей (bcrypt)
# или Argon2:
argon2-cffi==23.1.0

# --- CORS ---
Flask-CORS==4.0.0

# --- Environment Variables ---
python-dotenv==1.0.1

# --- Development Tools ---
black==24.10.0
isort==5.13.2
flake8==7.1.1
